Factorized resolvers and added a method to specify the encoding to be used by the reader.
