+++ src/main/java/com/schibsted/spt/data/jslt/impl/FileSystemResourceResolver.java
@@ -1,30 +1,15 @@
-
-// Copyright 2018 Schibsted Marketplaces Products & Technology As
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package com.schibsted.spt.data.jslt.impl;
 
-import com.schibsted.spt.data.jslt.JsltException;
-import com.schibsted.spt.data.jslt.ResourceResolver;
+import com.schibsted.spt.data.jslt.StreamResourceResolver;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.nio.file.*;
+import java.nio.file.FileSystems;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 
-public class FileSystemResourceResolver implements ResourceResolver {
+public final class FileSystemResourceResolver extends StreamResourceResolver {
 
     private final Path rootPath;
 
@@ -36,21 +21,14 @@ public class FileSystemResourceResolver implements ResourceResolver {
         this(FileSystems.getDefault().getPath("").toAbsolutePath());
     }
 
-    public Reader resolve(String jslt) {
-
+    @Override
+    protected InputStream getResourceAsStream(String jslt) throws IOException {
         Path jsltPath = FileSystems.getDefault().getPath(jslt).normalize();
         if (!jsltPath.isAbsolute()) {
-          jsltPath = rootPath.resolve(jsltPath).normalize();
+            jsltPath = rootPath.resolve(jsltPath).normalize();
         }
 
-        try {
-            final InputStream is = Files.newInputStream(jsltPath, StandardOpenOption.READ);
-            if (is == null) {
-                throw new JsltException("Cannot load '" + jslt + "' from file system: not found");
-            }
-            return new InputStreamReader(is, "utf-8");
-        } catch (IOException e) {
-            throw new JsltException("Couldn't load '" + jslt + "' from file system: " + e, e);
-        }
+        return Files.newInputStream(jsltPath, StandardOpenOption.READ);
     }
+
 }
